Several more UI changes, adds blocks to group slots.

+++ auctioneer/auction.py
@@ -1,13 +1,15 @@
-from datetime import datetime, timedelta
-
 from flask import Blueprint, flash, g, redirect, render_template, request, url_for
 from werkzeug.exceptions import abort
 
 from .auth import login_required
 from .constants import POSITIONS, TEAMS
 from .db import get_db
+from .utils import group_slots_by_block, query_range
 
 bp = Blueprint("auction", __name__)
+
+MINIMUM_BID_VALUE = 10
+MAX_NOMINATIONS_PER_BLOCK = 2
 
 BASE_QUERY = """
 SELECT nomination.id,
@@ -15,7 +17,9 @@ SELECT nomination.id,
     nomination.position,
     nomination.team,
     nomination.created_at,
+    slot.id as slot_id,
     slot.ends_at,
+    slot.block,
     nomination.nominator_id,
     user_n.username as nominator_username,
     nomination.matcher_id,
@@ -38,7 +42,9 @@ SELECT nomination.id,
     nomination.position,
     nomination.team,
     nomination.created_at,
+    slot.id as slot_id,
     slot.ends_at,
+    slot.block,
     nomination.nominator_id,
     user_n.username as nominator_username,
     nomination.matcher_id,
@@ -58,8 +64,6 @@ WHERE bid_u.user_id = ?
 ORDER by slot.ends_at DESC
 """
 
-MINIMUM_BID_VALUE = 10
-
 
 @bp.route("/")
 def index():
@@ -74,13 +78,10 @@ def index():
     closed_nominations = list()
     for nomination in nominations:
         if nomination["winner_id"]:
-            closed_nominations.append(nomination)
-        elif (
-            nomination["ends_at"]
-            and nomination["ends_at"] < datetime.utcnow()
-            and nomination["matcher_id"]
-        ):
-            match_nominations.append(nomination)
+            if nomination["matcher_id"]:
+                match_nominations.append(nomination)
+            else:
+                closed_nominations.append(nomination)
         else:
             open_nominations.append(nomination)
 
@@ -143,22 +144,27 @@ def nominate():
 
     db = get_db()
     users = db.execute("SELECT id, username from user").fetchall()
-    slots = get_slots()
+    slots = get_open_slots_for_user(days_range=(4, 10))
     if not slots:
-        flash(f"Maximum nominations reached for {g.user['username']}.")
+        flash(f"No nomination slots currently available for {g.user['username']}.")
+
+    blocks = group_slots_by_block(slots)
 
     return render_template(
         "auction/nominate.html",
         teams=TEAMS,
         positions=POSITIONS,
         users=users,
-        slots=slots,
+        blocks=blocks,
     )
 
 
 @bp.route("/<int:id>/update", methods=("GET", "POST"))
 @login_required
 def update(id):
+    if not g.user["is_league_manager"]:
+        abort(403)
+
     if request.method == "POST":
         name = request.form["name"]
         position = request.form["position"]
@@ -194,7 +200,13 @@ def update(id):
     db = get_db()
     nomination = get_nomination(id)
     users = db.execute("SELECT id, username from user").fetchall()
-    slots = get_slots(check_nominator=False)
+
+    slots = get_open_slots()
+    current_slot = db.execute(
+        "SELECT id, block, ends_at FROM slot WHERE id = ?", (nomination["slot_id"],)
+    ).fetchone()
+    slots.append(current_slot)
+    blocks = group_slots_by_block(slots)
 
     return render_template(
         "auction/update.html",
@@ -202,14 +214,14 @@ def update(id):
         positions=POSITIONS,
         teams=TEAMS,
         users=users,
-        slots=slots,
+        blocks=blocks,
     )
 
 
 @bp.route("/<int:id>/bid", methods=("GET", "POST"))
 @login_required
 def bid(id):
-    nomination = get_nomination(id, check_league_manager=False)
+    nomination = get_nomination(id)
     bid = dict(get_bid(g.user["id"], nomination["id"]))
     if not bid["value"]:
         bid["value"] = ""
@@ -241,9 +253,30 @@ def bid(id):
     return render_template("auction/bid.html", nomination=nomination, bid=bid)
 
 
+@bp.route("/<int:id>/match", methods=("GET", "POST"))
+@login_required
+def match(id):
+    nomination = get_nomination(id)
+
+    if request.method == "POST":
+        is_match = request.form["match"] == "yes"
+
+        db = get_db()
+        if is_match:
+            print("Match!")
+        else:
+            print("No match!")
+        return redirect(url_for("auction.index"))
+
+    return render_template("auction/match.html", nomination=nomination, bid=bid)
+
+
 @bp.route("/<int:id>/delete", methods=("POST",))
 @login_required
 def delete(id):
+    if not g.user["is_league_manager"]:
+        abort(403)
+
     get_nomination(id)
     db = get_db()
     db.execute("DELETE FROM nomination WHERE id = ?", (id,))
@@ -252,25 +285,37 @@ def delete(id):
     return redirect(url_for("auction.index"))
 
 
-def get_nomination(id, check_league_manager=True):
-    nomination = (
-        get_db()
-        .execute(
-            "SELECT n.id, name, position, team, slot_id, nominator_id, matcher_id, "
-            "winner_id, created_at, s.ends_at, username FROM nomination n "
-            "LEFT JOIN user u ON n.nominator_id = u.id "
-            "LEFT JOIN slot s ON n.slot_id = s.id "
-            "WHERE n.id = ?",
-            (id,),
-        )
-        .fetchone()
-    )
+NOMINATION_QUERY = """
+SELECT nomination.id,
+    nomination.name,
+    nomination.position,
+    nomination.team,
+    nomination.created_at,
+    slot.id as slot_id,
+    slot.ends_at,
+    slot.block,
+    nomination.nominator_id,
+    user_n.username as nominator_username,
+    nomination.matcher_id,
+    user_m.username as matcher_username,
+    nomination.winner_id,
+    user_w.username as winner_username,
+    bid_w.value as winner_value
+FROM nomination
+LEFT JOIN slot ON nomination.slot_id = slot.id
+LEFT JOIN user user_n ON nomination.nominator_id = user_n.id
+LEFT JOIN user user_m ON nomination.matcher_id = user_m.id
+LEFT JOIN user user_w ON nomination.winner_id = user_w.id
+LEFT JOIN bid bid_w ON nomination.id = bid_w.nomination_id AND nomination.winner_id = bid_w.user_id
+WHERE nomination.id = ?
+"""
+
+
+def get_nomination(id):
+    nomination = get_db().execute(NOMINATION_QUERY, (id,)).fetchone()
 
     if nomination is None:
         abort(404, f"Player id {id} doesn't exist.")
-
-    if check_league_manager and g.user["is_league_manager"] is True:
-        abort(403)
 
     return nomination
 
@@ -297,40 +342,42 @@ def get_bid(user_id, nomination_id):
     return bid
 
 
-def get_slots(
-    check_nominator=True,
-):
-    query_start = (datetime.utcnow() + timedelta(days=2)).strftime(
-        "%Y-%m-%d %H:%M:%S"
-    )
-    query_end = (datetime.utcnow() + timedelta(days=5)).strftime(
-        "%Y-%m-%d %H:%M:%S"
-    )
-
+def get_open_slots(days_range=None):
     db = get_db()
-    slots = db.execute(
-        "SELECT slot.id, ends_at FROM slot "
-        "LEFT JOIN nomination ON slot.id = nomination.slot_id "
-        "WHERE (nomination.slot_id IS NULL OR nomination.slot_id = slot.id) "
-        "AND ends_at BETWEEN ? AND ?",
-        (query_start, query_end),
-    ).fetchall()
-
-    if check_nominator:
-        user_nominations = db.execute(
-            "SELECT ends_at FROM slot "
+    if days_range:
+        slots = db.execute(
+            "SELECT slot.id, block, ends_at FROM slot "
             "LEFT JOIN nomination ON slot.id = nomination.slot_id "
-            "WHERE nomination.nominator_id = ? AND ends_at BETWEEN ? AND ?",
-            (g.user["id"], query_start, query_end),
+            "WHERE nomination.slot_id IS NULL "
+            "AND ends_at BETWEEN ? AND ?",
+            (query_range(days_range)),
         ).fetchall()
-        user_nomination_dates = [n["ends_at"].date() for n in user_nominations]
-
-        user_slots = list()
-        for slot in slots:
-            # TODO: Make this check work when all "blocks" are not on the same date
-            if slot["ends_at"].date() not in user_nomination_dates:
-                user_slots.append(slot)
-
-        slots = user_slots
+    else:
+        slots = db.execute(
+            "SELECT slot.id, block, ends_at FROM slot "
+            "LEFT JOIN nomination ON slot.id = nomination.slot_id "
+            "WHERE nomination.slot_id IS NULL "
+        ).fetchall()
 
     return slots
+
+
+def get_open_slots_for_user(days_range=(4, 10)):
+    slots = get_open_slots(days_range)
+
+    db = get_db()
+    user_nominations = db.execute(
+        "SELECT block, COUNT(*) as count FROM slot "
+        "LEFT JOIN nomination ON slot.id = nomination.slot_id "
+        "WHERE nomination.nominator_id = ? AND ends_at BETWEEN ? AND ? GROUP BY block",
+        (g.user["id"], *(query_range(days_range))),
+    ).fetchall()
+
+    user_nominations_per_block = {n["block"]: int(n["count"]) for n in user_nominations}
+
+    user_slots = list()
+    for slot in slots:
+        if user_nominations_per_block.get(slot["block"], 0) < MAX_NOMINATIONS_PER_BLOCK:
+            user_slots.append(slot)
+
+    return user_slots
